backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-multipart==0.0.6
pydantic==2.3.0
sqlalchemy==2.0.20
psycopg2-binary==2.9.7
pandas==2.1.0
pyyaml==6.0.1
python-dotenv==1.0.0